# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "# Notice: The URL in `Example Request` is currently wrong. The URL\
    \ to use is: `https://api.sandbox.mobilepay.dk/pos/api/v10/...`\n## PoS  V10 API\n\
    \nThe API is split into 5 categories: \n* **Brands**: Look up brands or stores\
    \ based on brand \n* **PointOfSales**: Management of Point of Sales including\
    \ create, delete and querying. Note that a Point of Sale is not editable. \n*\
    \ **Refunds**: Everything needed in order to do a refund (create, cancel and look\
    \ up) \n* **Stores**: Look up store information \n* **Payments**: Everything needed\
    \ in order to initiate and complete a payment. \n\n#### Payment flows\nPayments\
    \ can be started in two ways: as an initiated payment that is immediatedly ready\
    \ for user approval or as a prepared payment that is not yet ready for user approval\
    \ and must be explicitly marked as ready. Once a payment is ready for user approval,\
    \ it is issued to the user. Once the user accepts the payment request, and the\
    \ payment amount has been reserved, the payment must be captured to complete the\
    \ payment.\n\nThe API thus supports the following two payment flows:\n* Initiate\
    \ payment (POST /payments) --> poll until reserved (GET /payments/{paymentId})\
    \ --> capture (POST /payments/{paymentId}/capture) \n* Prepare payment (POST /payments/prepare)\
    \ --> mark as ready (POST /payments/{paymentId}/ready) --> poll until reserved\
...
basePath: "/pos"
schemes:
...
consumes:
...
produces:
...
security: []
paths:
...
  /api/v10/refunds/{refundId}:
    get:
      tags:
      - "Refunds"
      summary: "Lookup a Refund"
      operationId: "GetRefund"
      consumes: []
      produces:
      - "application/json"
      parameters:
      - name: "refundId"
        in: "path"
        description: "RefundId"
        required: true
        type: "string"
        format: "uuid"
      - name: "Authorization"
        in: "header"
        description: "Integrator's Bearer Token"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-Id"
        in: "header"
        description: "Integrator's MobilePay Client Id (Must be a valid GUID)"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-MobilePay-Client-System-Name"
        in: "header"
        description: "Integrator's Certified System Name"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-System-Version"
        in: "header"
        description: "Integrator's Certified System Version"
        required: true
        type: "string"
      responses:
        "200":
          description: "The Refund"
          examples:
            application/json:
              refundId: "6aee222f-703c-47f0-92ab-bc1c366b7d52"
              paymentId: "1cbfff94-3d17-4dc1-b667-5280e1ce50f9"
              refundOrderId: "REFUND-12345"
              amount: 12.5
              currencyCode: "DKK"
              status: "Reserved"
              pollDelayInMs: 100
          schema:
            $ref: "#/definitions/RefundResponse"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Refund not found"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Internal server error"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "Refunds"
  /api/v10/refunds:
    get:
      tags:
      - "Refunds"
      summary: "Lookup Refunds for a given filter"
      operationId: "QueryRefundIds"
      consumes: []
      produces:
      - "application/json"
      parameters:
      - name: "paymentId"
        in: "query"
        description: "Restricts Refunds returned to those that belongs to this paymentId"
        required: false
        type: "string"
        format: "uuid"
      - name: "Authorization"
        in: "header"
        description: "Integrator's Bearer Token"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-Id"
        in: "header"
        description: "Integrator's MobilePay Client Id (Must be a valid GUID)"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-MobilePay-Client-System-Name"
        in: "header"
        description: "Integrator's Certified System Name"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-System-Version"
        in: "header"
        description: "Integrator's Certified System Version"
        required: true
        type: "string"
      responses:
        "200":
          description: "List of Refunds"
          examples:
            application/json:
              refundIds:
              - "6aee222f-703c-47f0-92ab-bc1c366b7d52"
          schema:
            $ref: "#/definitions/QueryRefundIdsResponse"
        "204":
          description: "No Refunds found matching the filter"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "409":
          description: "Filter not specific enough"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Internal server error"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "Refunds"
    post:
      tags:
      - "Refunds"
      summary: "Create a refund (a payment can be refunded multiple times, the sum\
        \ of the refunds cannot exceed the payment full amount)"
      operationId: "CreateRefund"
      consumes:
      - "application/json-patch+json"
      - "application/json"
      - "text/json"
      - "application/*+json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "request"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/CreateRefundRequest"
      - name: "Authorization"
        in: "header"
        description: "Integrator's Bearer Token"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-Id"
        in: "header"
        description: "Integrator's MobilePay Client Id (Must be a valid GUID)"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-MobilePay-Client-System-Name"
        in: "header"
        description: "Integrator's Certified System Name"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-System-Version"
        in: "header"
        description: "Integrator's Certified System Version"
        required: true
        type: "string"
      - name: "X-MobilePay-Idempotency-Key"
        in: "header"
        description: "Used to ensure retried calls are handled correctly"
        required: true
        type: "string"
      responses:
        "200":
          description: "Refund has been created"
          examples:
            application/json:
              refundId: "6aee222f-703c-47f0-92ab-bc1c366b7d52"
          schema:
            $ref: "#/definitions/CreateRefundResponse"
        "400":
          description: "Bad Request. See error response for details"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Refund not found"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "409":
          description: "Refund could not be created"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Internal server error"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "Refunds"
  /api/v10/refunds/{refundId}/capture:
    post:
      tags:
      - "Refunds"
      summary: "Capture a refund"
      operationId: "CaptureRefund"
      consumes: []
      produces:
      - "application/json"
      parameters:
      - name: "refundId"
        in: "path"
        description: "Refund id"
        required: true
        type: "string"
        format: "uuid"
      - name: "Authorization"
        in: "header"
        description: "Integrator's Bearer Token"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-Id"
        in: "header"
        description: "Integrator's MobilePay Client Id (Must be a valid GUID)"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-MobilePay-Client-System-Name"
        in: "header"
        description: "Integrator's Certified System Name"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-System-Version"
        in: "header"
        description: "Integrator's Certified System Version"
        required: true
        type: "string"
      responses:
        "200":
          description: "Refund captured"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Refund not found"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "409":
          description: "Refund could not be captured"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Internal server error"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "Refunds"
  /api/v10/refunds/{refundId}/cancel:
    post:
      tags:
      - "Refunds"
      summary: "Cancel a refund"
      operationId: "CancelRefund"
      consumes: []
      produces:
      - "application/json"
      parameters:
      - name: "refundId"
        in: "path"
        description: "Refund id"
        required: true
        type: "string"
        format: "uuid"
      - name: "Authorization"
        in: "header"
        description: "Integrator's Bearer Token"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-Id"
        in: "header"
        description: "Integrator's MobilePay Client Id (Must be a valid GUID)"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-MobilePay-Client-System-Name"
        in: "header"
        description: "Integrator's Certified System Name"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-System-Version"
        in: "header"
        description: "Integrator's Certified System Version"
        required: true
        type: "string"
      responses:
        "200":
          description: "Refund cancelled"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Refund could not be cancelled"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "409":
          description: "Conflict"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Internal server error"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "Refunds"
  /api/v10/stores/{storeId}:
    get:
      tags:
      - "Stores"
      summary: "Lookup a Store"
      operationId: "GetStore"
      consumes: []
      produces:
      - "application/json"
      parameters:
      - name: "storeId"
        in: "path"
        description: "StoreId"
        required: true
        type: "string"
        format: "uuid"
      - name: "Authorization"
        in: "header"
        description: "Integrator's Bearer Token"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-Id"
        in: "header"
        description: "Integrator's MobilePay Client Id (Must be a valid GUID)"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-MobilePay-Client-System-Name"
        in: "header"
        description: "Integrator's Certified System Name"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-System-Version"
        in: "header"
        description: "Integrator's Certified System Version"
        required: true
        type: "string"
      responses:
        "200":
          description: "The Store"
          examples:
            application/json:
              storeName: "Store 1"
              storeId: "939f2b0d-152e-4000-8e04-2e30aa9faf7f"
              merchantBrandId: "POSDK12345"
              merchantLocationId: "10001"
          schema:
            $ref: "#/definitions/StoreResponse"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Store not found"
      x-swagger-router-controller: "Stores"
  /api/v10/stores:
    get:
      tags:
      - "Stores"
      summary: "Lookup Stores for a given filter"
      operationId: "GetStores"
      consumes: []
      produces:
      - "application/json"
      parameters:
      - name: "merchantBrandId"
        in: "query"
        description: "MerchantBrandId"
        required: false
        type: "string"
      - name: "merchantLocationId"
        in: "query"
        description: "MerchantLocationId"
        required: false
        type: "string"
      - name: "Authorization"
        in: "header"
        description: "Integrator's Bearer Token"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-Id"
        in: "header"
        description: "Integrator's MobilePay Client Id (Must be a valid GUID)"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-MobilePay-Client-System-Name"
        in: "header"
        description: "Integrator's Certified System Name"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-System-Version"
        in: "header"
        description: "Integrator's Certified System Version"
        required: true
        type: "string"
      responses:
        "200":
          description: "A list of Stores"
          examples:
            application/json:
              storeIds:
              - "939f2b0d-152e-4000-8e04-2e30aa9faf7f"
          schema:
            $ref: "#/definitions/StoreIdsResponse"
        "400":
          description: "Bad Request. See error response for details"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Stores not found"
      x-swagger-router-controller: "Stores"
definitions:
...
  ErrorResponse:
    type: "object"
    properties:
      code:
        type: "string"
        readOnly: true
      message:
        type: "string"
        readOnly: true
      correlationId:
        type: "string"
        readOnly: true
...
  RefundResponse:
    type: "object"
    required:
    - "amount"
    - "currencyCode"
    - "paymentId"
    - "pollDelayInMs"
    - "refundId"
    - "refundOrderId"
    - "status"
    properties:
      refundId:
        type: "string"
        format: "uuid"
      paymentId:
        type: "string"
        format: "uuid"
      refundOrderId:
        type: "string"
      amount:
        type: "number"
        format: "double"
      currencyCode:
        type: "string"
        enum:
        - "DKK"
        - "EUR"
      status:
        type: "string"
        enum:
        - "Initiated"
        - "Reserved"
        - "Captured"
        - "CancelledByMerchant"
        - "CancelledByMobilePay"
      pollDelayInMs:
        type: "integer"
        format: "int32"
  QueryRefundIdsResponse:
    type: "object"
    required:
    - "refundIds"
    properties:
      refundIds:
        type: "array"
        items:
          type: "string"
          format: "uuid"
  CreateRefundRequest:
    type: "object"
    required:
    - "amount"
    - "currencyCode"
    - "paymentId"
    - "refundOrderId"
    properties:
      paymentId:
        type: "string"
        format: "uuid"
        description: "PaymentId of the payment to refund"
      refundOrderId:
        type: "string"
      amount:
        type: "number"
        format: "double"
        description: "Amount to refund (Cannot be more than the Payment's amount)"
      currencyCode:
        type: "string"
        enum:
        - "DKK"
        - "EUR"
    example:
      paymentId: "1cbfff94-3d17-4dc1-b667-5280e1ce50f9"
      refundOrderId: "REFUND-12345"
      amount: 12.5
      currencyCode: "DKK"
  CreateRefundResponse:
    type: "object"
    required:
    - "refundId"
    properties:
      refundId:
        type: "string"
        format: "uuid"
  StoreResponse:
    type: "object"
    required:
    - "merchantBrandId"
    - "merchantLocationId"
    - "storeId"
    - "storeName"
    properties:
      storeName:
        type: "string"
        description: "Store Name"
      storeId:
        type: "string"
        format: "uuid"
        description: "StoreId"
      merchantBrandId:
        type: "string"
      merchantLocationId:
        type: "string"
  StoreIdsResponse:
    type: "object"
    required:
    - "storeIds"
    properties:
      storeIds:
        type: "array"
        items:
          type: "string"
          format: "uuid"
...